yacht integer values

+++ src/put/sailhero/android/utils/UserProfileResponseCreator.java
@@ -35,8 +35,8 @@ public class UserProfileResponseCreator implements ResponseCreator<UserProfileRe
 					Yacht yacht = new Yacht();
 					yacht.setId(Integer.valueOf(yachtObject.get("id").toString()));
 					yacht.setName(yachtObject.get("name").toString());
-					yacht.setLength(Float.valueOf(yachtObject.get("length").toString()));
-					yacht.setWidth(Float.valueOf(yachtObject.get("width").toString()));
+					yacht.setLength(Integer.valueOf(yachtObject.get("length").toString()));
+					yacht.setWidth(Integer.valueOf(yachtObject.get("width").toString()));
 					yacht.setCrew(Integer.valueOf(yachtObject.get("crew").toString()));
 
 					userProfileResponse.setYacht(yacht);
